requirements.txt
treamlit
pandas
plotly-express
openpyxl